added folder for each componenet, changed static file data into actual comment componenet with props, organized the code/folder structure
