--- src/components/post/index.js
@@ -1,1 +1,0 @@
-x
