+++ src/views/Home.vue
@@ -19,7 +19,7 @@
 
 
 <script>
-import post from "@/components/post.vue";
+import post from "@/components/post/post.vue";
 import '../styles/posts.scss'
 import '../styles/home.scss'
