+++ src/views/comments.vue
@@ -1,24 +1,14 @@
 <template>
 <div id="home">
      <div class="wrapper"v-for="comment in comments" >
-      <div class="comment col-sm-4"  >
+    <comment :name="comment.name" :email="comment.email" :body="comment.body" />   
 
-        <div id="title">
-          {{comment.name}}
-          </div>
-        <div class="email">
-        {{comment.email}}
-        </div>
-          <p>{{comment.body}}</p>
-        </div>
-    
-      </div>
       </div>
       </div>
 </template>
 
 <script>
-import comment from "@/components/comment.vue";
+import comment from "@/components/comment/comment.vue";
 import '../styles/postcomments.scss'
 
 const fetch = window.location.pathname.split("/")[2];
