+++ src/views/postcomments.vue
@@ -1,19 +1,9 @@
 <template>
 <div id="home">
      <div class="wrapper"v-for="comment in commentList" >
-      <div class="comment col-sm-4"  >
-
-        <div id="title">
-          {{comment.name}}
-          </div>
-        <div class="email">
-        {{comment.email}}
-        </div>
-          <p>{{comment.body}}</p>
-        </div>
     
-      </div>
-      </div>
+ <comment :name="comment.name" :email="comment.email" :body="comment.body" />   
+   </div>
       </div>
 </template>
 
@@ -21,8 +11,9 @@
 
 
 <script>
-import comment from "@/components/comment.vue";
+import comment from "@/components/comment/comment.vue";
 import '../styles/postcomments.scss'
+import { maxHeaderSize } from 'http';
 const fetch = window.location.pathname.split(":")[1];
 export default {
 
@@ -54,6 +45,11 @@ export default {
   name: "postcomments",
   components: {
     comment
+  },
+   props: {
+    name: String,
+    email: String,
+    body: String
   }
 };
 </script>
